fix:结构调整后的异常状况
