+++ forum-web/src/main/java/com/github/liuyueyi/forum/web/front/article/rest/ArticleRestController.java
@@ -4,6 +4,7 @@ import com.github.liueyueyi.forum.api.model.context.ReqInfoContext;
 import com.github.liueyueyi.forum.api.model.enums.DocumentTypeEnum;
 import com.github.liueyueyi.forum.api.model.enums.OperateTypeEnum;
 import com.github.liueyueyi.forum.api.model.vo.ResVo;
+import com.github.liueyueyi.forum.api.model.vo.article.ArticlePostReq;
 import com.github.liueyueyi.forum.api.model.vo.article.dto.CategoryDTO;
 import com.github.liueyueyi.forum.api.model.vo.article.dto.TagDTO;
 import com.github.liueyueyi.forum.api.model.vo.constants.StatusEnum;
@@ -11,12 +12,16 @@ import com.github.liuyueyi.forum.core.permission.Permission;
 import com.github.liuyueyi.forum.core.permission.UserRole;
 import com.github.liuyueyi.forum.service.article.repository.entity.ArticleDO;
 import com.github.liuyueyi.forum.service.article.service.ArticleReadService;
+import com.github.liuyueyi.forum.service.article.service.ArticleWriteService;
 import com.github.liuyueyi.forum.service.article.service.CategoryService;
 import com.github.liuyueyi.forum.service.article.service.TagService;
 import com.github.liuyueyi.forum.service.user.service.UserFootService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import java.util.List;
 
 /**
@@ -36,6 +41,8 @@ public class ArticleRestController {
     private CategoryService categoryService;
     @Autowired
     private TagService tagService;
+    @Autowired
+    private ArticleWriteService articleWriteService;
 
 
     /**
@@ -95,5 +102,25 @@ public class ArticleRestController {
                 ReqInfoContext.getReqInfo().getUserId(),
                 operate);
         return ResVo.ok(true);
+    }
+
+
+    /**
+     * 发布文章，完成后跳转到详情页
+     * - 这里有一个重定向的知识点
+     * - fixme 博文：* [5.请求重定向 | 一灰灰Learning](https://hhui.top/spring-web/02.response/05.190929-springboot%E7%B3%BB%E5%88%97%E6%95%99%E7%A8%8Bweb%E7%AF%87%E4%B9%8B%E9%87%8D%E5%AE%9A%E5%90%91/)
+     *
+     * @return
+     */
+    @Permission(role = UserRole.LOGIN)
+    @PostMapping(path = "post")
+    @ResponseBody
+    @Transactional(rollbackFor = Exception.class)
+    public ResVo<Long> post(@RequestBody ArticlePostReq req, HttpServletResponse response) throws IOException {
+        Long id = articleWriteService.saveArticle(req, ReqInfoContext.getReqInfo().getUserId());
+//        return "redirect:/article/detail/" + id;
+//        response.sendRedirect("/article/detail/" + id);
+        // 这里采用前端重定向策略
+        return ResVo.ok(id);
     }
 }
